fix: implemented useEffect on gameSocket and gamerendering
